+++ general/file.h
@@ -12,7 +12,7 @@ const char* const FUNCTION_FAILURE_ALERT = "failed";
 struct DataFromInputFIle {
     size_t fileSize;
     FILE* inputFile;
-    int nStrings;
+    size_t nStrings;
     char* buffer;
     size_t bufferSize;
     string* strings;
